Add calculator functionality
The general functionality for the calculator has the user click on the
keys in order to activiate event listeners. The event listeners add the
numbers of the keys to the screen and once the user clicks on an operator
that number is saved into a varible and the specific operator is saved.
The user can then select more number keys, which will replace the
previous number on the screen. A boolean is used in this case to detect
if an operator was just selected or not. If so, the new number will
replace the current number on the screen, otherwise, it will be tacked
onto it. When the equals operator was clicked, the second number
would be stored in another variable which would be passed into the
operate function which would call the corresonding operation function
(add, subtract, etc).

In order to prevent the screen from overflowing, the string length for
the number box was limited to 8. In order to accomodate for decimals,
the method toFixed() was used if the resulting number had a decimal after
it and would round the decimal to fit the screen.